+++ CHANGELOG.md
@@ -9,6 +9,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added support for `:hover` pseudo-class.
+- Added support for [Pseudo-classes](https://developer.mozilla.org/en-US/docs/Web/CSS/Pseudo-classes).
+
+## [0.5.1]
+
+### Added
+
 - New selectors for setting up BorderColor and Image texture.
 - New documentation available on: [link](https://afonsolage.github.io/bevy_ecss/).